backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
numpy
opencv-python-headless
mediapipe
librosa
moviepy
openai-whisper
groq